Move formatter out of the controller
