+++ MVC/Ponto/Model/CommonFormatters.swift
@@ -1,0 +1,24 @@
+//
+//  CommonFormatters.swift
+//  Ponto-MVC
+//
+//  Created by Carolina Cruz Agra Lopes on 22/05/22.
+//
+
+import Foundation
+
+class CommonFormatters {
+
+    let dateComponentsFormatter: DateComponentsFormatter = {
+        let formatter = DateComponentsFormatter()
+        formatter.unitsStyle = .positional
+        formatter.zeroFormattingBehavior = .pad
+        formatter.allowedUnits = [.hour, .minute, .second]
+        return formatter
+    }()
+
+    static let shared: CommonFormatters = CommonFormatters()
+
+    private init() { }
+
+}
